// client2/src/types/Athelete.ts
export interface AthleteLevel {
  id?: number;
  name: string;
  number: number;
  description?: string;
}

export interface Athlete {
  id?: number;
  firstname: string;
  lastname: string;
  name: string;
  level: AthleteLevel | number;
  patronymic: string;
  birth_year: string;
  picture: string;
  sport_type: string;
  created_at: string;
  updated_at: string;
}

export interface AthleteParams {
  id?: number;
  athlete: Athlete | number;
  bmi: number;
  weight: number;
  height: number;
  created_at: string;
  description?: string;
}